feat: implement queries and formatter in separate files
