+++ src/api/news/queries.ts
@@ -1,0 +1,9 @@
+import { useQuery } from "@tanstack/react-query";
+import { getNewsBySection } from "./module";
+
+export const useNews = (sectionName: string) => {
+  return useQuery({
+    queryKey: ["section", sectionName],
+    queryFn: () => getNewsBySection(sectionName),
+  });
+};
